+++ bootstrap.sh
@@ -11,7 +11,6 @@ if [ -x "$(which python)" ]; then
     ./py2/bin/pip install repoze.zcml==0.4
     ./py2/bin/pip install repoze.workflow==0.6.1
     ./py2/bin/pip install https://github.com/bluedynamics/node.ext.ugm/archive/master.zip
-    ./py2/bin/pip install https://github.com/bluedynamics/node.ext.zodb/archive/master.zip
     ./py2/bin/pip install https://github.com/bluedynamics/yafowil/archive/master.zip
     ./py2/bin/pip install https://github.com/bluedynamics/cone.tile/archive/master.zip
     ./py2/bin/pip install https://github.com/bluedynamics/cone.app/archive/master.zip
@@ -28,7 +27,6 @@ if [ -x "$(which python3)" ]; then
     ./py3/bin/pip install repoze.zcml==1.0b1
     ./py3/bin/pip install repoze.workflow==1.0b1
     ./py3/bin/pip install https://github.com/bluedynamics/node.ext.ugm/archive/master.zip
-    ./py3/bin/pip install https://github.com/bluedynamics/node.ext.zodb/archive/master.zip
     ./py3/bin/pip install https://github.com/bluedynamics/yafowil/archive/master.zip
     ./py3/bin/pip install https://github.com/bluedynamics/cone.tile/archive/master.zip
     ./py3/bin/pip install https://github.com/bluedynamics/cone.app/archive/master.zip