+++ CHANGES.rst
@@ -3,7 +3,22 @@ Changes
 =======
 
 
-1.0b1 (unreleased)
+1.0a1 (unreleased)
 ------------------
 
-- no changes yet.
+- Set ``node.interfaces.IOrdered`` on ``cone.zodb.entry.ZODBEntry`` to fix
+  ``treerepr``.
+  [rnix]
+
+- Python 3 compatibility.
+  [rnix]
+
+- Upgrade to ``cone.app`` 1.0b1.
+  [rnix]
+
+
+< 1.0
+-----
+
+- Initial work.
+  [rnix]